new travelPlanning will be saved in db
