+++ Travel/View/PhotoAlbumCollectionViewCell.swift
